backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.0
aiosqlite>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=5.0.0
celery[redis]>=5.3
google-cloud-documentai>=2.20
google-cloud-storage>=2.10
msal>=1.26
pywebpush>=2.0
jinja2>=3.1.0
weasyprint>=60.0
